chore: Move fit image to screen button
