+++ web_ui/src/pages/annotator/components/fit-image-to-screen-button/fit-image-to-screen-button.test.tsx
